+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-
+-  Encrypt all messages and logs
 -  Refactored the message schema to be simpler
 
 ## [3.1.4] - 2023-04-10